+++ models/PunchItem.go
@@ -8,7 +8,7 @@ import (
 
 // PunchItem 打卡事项
 type PunchItem struct {
-	ID             int          `orm:"column(id);auto;pk"`
+	ID             int64        `orm:"column(id);auto;pk"`
 	Title          string       `orm:"size(50)"`                        // 打卡事项标题
 	Content        string       `orm:"size(500)"`                       // 打卡事项正文
 	Creator        *User        `orm:"rel(fk)"`                         // 创建人；当某个用户被注销时，此项可能为空，因此用户注销时，一般只是设置其状态为“注销”状态，否则此处为 null 易出问题。
@@ -23,7 +23,7 @@ type PunchItem struct {
 	EndTime        time.Time    `orm:"type(datetime)"`                  // 结束打卡时间
 	PeriodUnit     int          `orm:"size(5);default(3)"`              // 周期的单位：0：秒，1：分钟；2：小时；3.天；4：周；5：月；6：季度；7：半年；8：年
 	PeriodValue    int64        `orm:"default(1)"`                      // 打卡周期，默认一天周期
-	ActiveBonus    int          `orm:"default(0)"`                      // 活跃积分（活跃度）
+	ActiveBonus    int64        `orm:"default(0)"`                      // 活跃积分（活跃度）
 	Groups         []*Group     `orm:"rel(m2m);null"`                   // 归属于群
 	Punchers       []*User      `orm:"reverse(many);null"`              // 关联的打卡人
 	PunchRecord    *PunchRecord `orm:"reverse(one);null"`               // 一个打卡事项对应一个打卡记录
